Replace attach() / detach() pair with envir= named argument to eval()
Instead of using the attach() / detach() pair we pass the data frame
containing the variables as the envir argument to eval. This works
because eval will treat anything remotely resembling a list as a valid
environment. data does not have a parent environment, so we need to
match all variables explicitly and include them in the data.frame data.